security toegevoegd aan requests in program.cs

+++ Program.cs
@@ -51,29 +51,23 @@ app.MapGet("/", () => "Hello World!");
 
 // GET
 
-app.MapGet("/leden", async (ILidService lidService) => await lidService.GetAllLeden());
+app.MapGet("/leden", [Authorize] async (ILidService lidService) => await lidService.GetAllLeden());
 
-// app.MapGet("/leden", [Authorize] async (ILidService lidService, ClaimsPrincipal user) => {
-//     return Results.Ok(await lidService.GetAllLeden());
-// });
+app.MapGet("/lid/{lidId}", [Authorize] async (ILidService lidService, string lidId) => await lidService.GetLid(lidId));
 
-app.MapGet("/lid/{lidId}", async (ILidService lidService, string lidId) => await lidService.GetLid(lidId));
+app.MapGet("/leden/tak/{takId}", [Authorize] async (ILidService lidService, string takId) => await lidService.GetLedenByTakId(takId));
 
-app.MapGet("/leden/tak/{takId}", async (ILidService lidService, string takId) => await lidService.GetLedenByTakId(takId));
+app.MapGet("/leden/groep/{groepId}", [Authorize] async (ILidService lidService, string groepId) => await lidService.GetLedenByGroepId(groepId));
 
-app.MapGet("/leden/groep/{groepId}", async (ILidService lidService, string groepId) => await lidService.GetLedenByGroepId(groepId));
+app.MapGet("/takken", [Authorize] async (ILidService lidService) => await lidService.GetAllTakken());
 
-app.MapGet("/takken", async (ILidService lidService) => await lidService.GetAllTakken());
+app.MapGet("/tak/{takId}", [Authorize] async (ILidService lidService, string takId) => await lidService.GetTak(takId));
 
-app.MapGet("/tak/{takId}", async (ILidService lidService, string takId) => await lidService.GetTak(takId));
+app.MapGet("/groepen", [Authorize] async (ILidService lidService) => await lidService.GetAllGroepen());
 
-app.MapGet("/groepen", async (ILidService lidService) => await lidService.GetAllGroepen());
+app.MapGet("/groep/{groepId}", [Authorize] async (ILidService lidService, string groepId) => await lidService.GetGroep(groepId));
 
-app.MapGet("/groep/{groepId}", async (ILidService lidService, string groepId) => await lidService.GetGroep(groepId));
-
-// POST
-
-app.MapPost("/lid", async (ILidService lidService, IValidator<Lid> validator, Lid lid) => {
+app.MapPost("/lid", [Authorize] async (ILidService lidService, IValidator<Lid> validator, Lid lid, ClaimsPrincipal user) => {
     var validatorResult = validator.Validate(lid);
     if (validatorResult.IsValid){
         var result = await lidService.AddLid(lid);
@@ -84,7 +78,7 @@ app.MapPost("/lid", async (ILidService lidService, IValidator<Lid> validator, Li
     }
 });
 
-app.MapPost("/tak", async (ILidService lidService, IValidator<Tak> validator, Tak tak) => {
+app.MapPost("/tak", [Authorize] async (ILidService lidService, IValidator<Tak> validator, Tak tak) => {
     var validatorResult = validator.Validate(tak);
     if (validatorResult.IsValid){
         var result = await lidService.AddTak(tak);
@@ -95,7 +89,7 @@ app.MapPost("/tak", async (ILidService lidService, IValidator<Tak> validator, Ta
     }
 });
 
-app.MapPost("/groep", async (ILidService lidService, IValidator<Groep> validator, Groep groep) => {
+app.MapPost("/groep", [Authorize] async (ILidService lidService, IValidator<Groep> validator, Groep groep) => {
     var validatorResult = validator.Validate(groep);
     if (validatorResult.IsValid){
         var result = await lidService.AddGroep(groep);
@@ -109,19 +103,19 @@ app.MapPost("/groep", async (ILidService lidService, IValidator<Groep> validator
 
 // PUT
 
-app.MapPut("/lid/{lidId}", async (ILidService lidService, string lidId, Lid lid) => await lidService.UpdateLid(lidId, lid));
+app.MapPut("/lid/{lidId}", [Authorize] async (ILidService lidService, string lidId, Lid lid) => await lidService.UpdateLid(lidId, lid));
 
-app.MapPut("/tak/{takId}", async (ILidService lidService, string takId, Tak tak) => await lidService.UpdateTak(takId, tak));
+app.MapPut("/tak/{takId}", [Authorize] async (ILidService lidService, string takId, Tak tak) => await lidService.UpdateTak(takId, tak));
 
-app.MapPut("/groep/{groepId}", async (ILidService lidService, string groepId, Groep groep) => await lidService.UpdateGroep(groepId, groep));
+app.MapPut("/groep/{groepId}", [Authorize] async (ILidService lidService, string groepId, Groep groep) => await lidService.UpdateGroep(groepId, groep));
 
 // DELETE
 
-app.MapDelete("/lid/{lidId}", async (ILidService lidService, string lidId) => await lidService.DeleteLid(lidId));
+app.MapDelete("/lid/{lidId}", [Authorize] async (ILidService lidService, string lidId) => await lidService.DeleteLid(lidId));
 
-app.MapDelete("/tak/{takId}", async (ILidService lidService, string takId) => await lidService.DeleteLid(takId));
+app.MapDelete("/tak/{takId}", [Authorize] async (ILidService lidService, string takId) => await lidService.DeleteLid(takId));
 
-app.MapDelete("/groep/{groepId}", async (ILidService lidService, string groepId) => await lidService.DeleteLid(groepId));
+app.MapDelete("/groep/{groepId}", [Authorize] async (ILidService lidService, string groepId) => await lidService.DeleteLid(groepId));
 
 // AUTHENTICATION
